refactored code to make validation more modular. Added CustomerController
