+++ BangazonTerminalInterface/BangazonTerminalInterface.csproj
@@ -44,6 +44,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Controllers\CustomerController.cs" />
     <Compile Include="DAL\Repository\CartRepository.cs" />
     <Compile Include="DAL\Repository\CustomerRepository.cs" />
     <Compile Include="DAL\Repository\PaymentRepository.cs" />
